% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raquifer_package.R, R/aquifer_interface.R
\docType{package}
\name{Raquifer-package}
\alias{Raquifer}
\alias{Raquifer-package}
\title{Raquifer: Estimate the Water Influx into Hydrocarbon Reservoirs}
\description{
Generate a table of cumulative water influx into hydrocarbon reservoirs over time using un-steady and pseudo-steady state models. Van Everdingen, A. F. and Hurst, W. (1949) <doi:10.2118/949305-G>. Fetkovich, M. J. (1971) <doi:10.2118/2603-PA>. Yildiz, T. and Khosravi, A. (2007) <doi:10.2118/103283-PA>.

Generate a table of cumulative water influx into hydrocarbon reservoirs over time using un-steady and pseudo-steady state models. Van Everdingen, A. F. and Hurst, W. (1949) <doi:10.2118/949305-G>. Fetkovich, M. J. (1971) <doi:10.2118/2603-PA>. Yildiz, T. and Khosravi, A. (2007) <doi:10.2118/103283-PA>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://susaenergy.github.io/Raquifer_ws/}
}


Useful links:
\itemize{
  \item \url{https://susaenergy.github.io/Raquifer_ws/}
}

}
\author{
\strong{Maintainer}: Farshad Tabasinejad \email{farshad.tabasinejad@susaenergy.com}


\strong{Maintainer}: Farshad Tabasinejad \email{farshad.tabasinejad@susaenergy.com}

}
\keyword{internal}
